separate jobmaker code per blockchain type
